[backend] fix test (#issue/3166)
