--- opencti-platform/opencti-graphql/tests/02-integration/02-resolvers/threatActorGroup-test.js
@@ -1,220 +1,0 @@
-import { expect, it, describe } from 'vitest';
-import gql from 'graphql-tag';
-import { queryAsAdmin } from '../../utils/testQuery';
-
-const LIST_QUERY = gql`
-  query threatActorsGroup(
-    $first: Int
-    $after: ID
-    $orderBy: threatActorsGroupOrdering
-    $orderMode: OrderingMode
-    $filters: [threatActorsGroupFiltering]
-    $filterMode: FilterMode
-    $search: String
-  ) {
-    threatActorsGroup(
-      first: $first
-      after: $after
-      orderBy: $orderBy
-      orderMode: $orderMode
-      filters: $filters
-      filterMode: $filterMode
-      search: $search
-    ) {
-      edges {
-        node {
-          id
-          name
-          description
-        }
-      }
-    }
-  }
-`;
-
-const READ_QUERY = gql`
-  query threatActorGroup($id: String!) {
-    threatActorGroup(id: $id) {
-      id
-      name
-      description
-      toStix
-    }
-  }
-`;
-
-describe('Threat actor resolver standard behavior', () => {
-  let threatActorGroupInternalId;
-  const threatActorsGroupStixId = 'threat-actor-group--16978493-d5fb-4b28-a39a-eca332f53189';
-  it('should threat actor created', async () => {
-    const CREATE_QUERY = gql`
-      mutation threatActorGroupAdd($input: ThreatActorGroupAddInput!) {
-        threatActorGroupAdd(input: $input) {
-          id
-          name
-          description
-        }
-      }
-    `;
-    // Create the threat actor
-    const THREAT_ACTOR_TO_CREATE = {
-      input: {
-        name: 'Threat actor group',
-        stix_id: threatActorsGroupStixId,
-        description: 'Threat actor description',
-      },
-    };
-    const threatActorGroup = await queryAsAdmin({
-      query: CREATE_QUERY,
-      variables: THREAT_ACTOR_TO_CREATE,
-    });
-    expect(threatActorGroup).not.toBeNull();
-    expect(threatActorGroup.data.threatActorGroupAdd).not.toBeNull();
-    expect(threatActorGroup.data.threatActorGroupAdd.name).toEqual('Threat actor group');
-    threatActorGroupInternalId = threatActorGroup.data.threatActorGroupAdd.id;
-  });
-  it('should threat actor loaded by internal id', async () => {
-    const queryResult = await queryAsAdmin({ query: READ_QUERY, variables: { id: threatActorGroupInternalId } });
-    expect(queryResult).not.toBeNull();
-    expect(queryResult.data.threatActorGroup).not.toBeNull();
-    expect(queryResult.data.threatActorGroup.id).toEqual(threatActorGroupInternalId);
-    expect(queryResult.data.threatActorGroup.toStix.length).toBeGreaterThan(5);
-  });
-  it('should threat actor loaded by stix id', async () => {
-    const queryResult = await queryAsAdmin({ query: READ_QUERY, variables: { id: threatActorsGroupStixId } });
-    expect(queryResult).not.toBeNull();
-    expect(queryResult.data.threatActorGroup).not.toBeNull();
-    expect(queryResult.data.threatActorGroup.id).toEqual(threatActorGroupInternalId);
-  });
-  it('should list threat actors', async () => {
-    const queryResult = await queryAsAdmin({ query: LIST_QUERY, variables: { first: 1 } });
-    expect(queryResult.data.threatActorsGroup.edges.length).toEqual(1);
-  });
-  it('should update threat actor', async () => {
-    const UPDATE_QUERY = gql`
-      mutation threatActorGroupEdit($id: ID!, $input: [EditInput]!) {
-        threatActorGroupEdit(id: $id) {
-          fieldPatch(input: $input) {
-            id
-            name
-          }
-        }
-      }
-    `;
-    const queryResult = await queryAsAdmin({
-      query: UPDATE_QUERY,
-      variables: { id: threatActorGroupInternalId, input: { key: 'name', value: ['Threat actor - test'] } },
-    });
-    expect(queryResult.data.threatActorGroupEdit.fieldPatch.name).toEqual('Threat actor - test');
-  });
-  it('should context patch threat actor', async () => {
-    const CONTEXT_PATCH_QUERY = gql`
-      mutation threatActorGroupEdit($id: ID!, $input: EditContext) {
-        threatActorGroupEdit(id: $id) {
-          contextPatch(input: $input) {
-            id
-          }
-        }
-      }
-    `;
-    const queryResult = await queryAsAdmin({
-      query: CONTEXT_PATCH_QUERY,
-      variables: { id: threatActorGroupInternalId, input: { focusOn: 'description' } },
-    });
-    expect(queryResult.data.threatActorGroupEdit.contextPatch.id).toEqual(threatActorGroupInternalId);
-  });
-  it('should context clean threat actor', async () => {
-    const CONTEXT_PATCH_QUERY = gql`
-      mutation threatActorGroupEdit($id: ID!) {
-        threatActorGroupEdit(id: $id) {
-          contextClean {
-            id
-          }
-        }
-      }
-    `;
-    const queryResult = await queryAsAdmin({
-      query: CONTEXT_PATCH_QUERY,
-      variables: { id: threatActorGroupInternalId },
-    });
-    expect(queryResult.data.threatActorGroupEdit.contextClean.id).toEqual(threatActorGroupInternalId);
-  });
-  it('should add relation in threat actor', async () => {
-    const RELATION_ADD_QUERY = gql`
-      mutation threatActorGroupEdit($id: ID!, $input: StixRefRelationshipAddInput!) {
-        threatActorGroupEdit(id: $id) {
-          relationAdd(input: $input) {
-            id
-            from {
-              ... on ThreatActorGroup {
-                objectMarking {
-                  edges {
-                    node {
-                      id
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    `;
-    const queryResult = await queryAsAdmin({
-      query: RELATION_ADD_QUERY,
-      variables: {
-        id: threatActorGroupInternalId,
-        input: {
-          toId: 'marking-definition--78ca4366-f5b8-4764-83f7-34ce38198e27',
-          relationship_type: 'object-marking',
-        },
-      },
-    });
-    expect(queryResult.data.threatActorGroupEdit.relationAdd.from.objectMarking.edges.length).toEqual(1);
-  });
-  it('should delete relation in threat actor', async () => {
-    const RELATION_DELETE_QUERY = gql`
-      mutation threatActorGroupEdit($id: ID!, $toId: StixRef!, $relationship_type: String!) {
-        threatActorGroupEdit(id: $id) {
-          relationDelete(toId: $toId, relationship_type: $relationship_type) {
-            id
-            objectMarking {
-              edges {
-                node {
-                  id
-                }
-              }
-            }
-          }
-        }
-      }
-    `;
-    const queryResult = await queryAsAdmin({
-      query: RELATION_DELETE_QUERY,
-      variables: {
-        id: threatActorGroupInternalId,
-        toId: 'marking-definition--78ca4366-f5b8-4764-83f7-34ce38198e27',
-        relationship_type: 'object-marking',
-      },
-    });
-    expect(queryResult.data.threatActorGroupEdit.relationDelete.objectMarking.edges.length).toEqual(0);
-  });
-  it('should threat actor deleted', async () => {
-    const DELETE_QUERY = gql`
-      mutation threatActorGroupDelete($id: ID!) {
-        threatActorGroupEdit(id: $id) {
-          delete
-        }
-      }
-    `;
-    // Delete the threat actor
-    await queryAsAdmin({
-      query: DELETE_QUERY,
-      variables: { id: threatActorGroupInternalId },
-    });
-    // Verify is no longer found
-    const queryResult = await queryAsAdmin({ query: READ_QUERY, variables: { id: threatActorsGroupStixId } });
-    expect(queryResult).not.toBeNull();
-    expect(queryResult.data.threatActorGroup).toBeNull();
-  });
-});
